REFACTOR: componentmax Entry.tsx into smaller components
