+++ app/merch/_components/entry/EntrySubtitle.tsx
@@ -1,0 +1,5 @@
+export default function EntrySubtitle() {
+    return (
+        <p className="text-white text-2xl md:text-5xl mt-4">Tap to Enter.</p>
+    );
+}
